Refactor gallery deprecations
