+++ packages/block-library/src/gallery/deprecated/index.js
@@ -1,0 +1,13 @@
+/**
+ * Internal dependencies
+ */
+import v1 from './v1';
+import v2 from './v2';
+import v3 from './v3';
+import v4 from './v4';
+import v5 from './v5';
+import v6 from './v6';
+
+const deprecated = [ v1, v2, v3, v4, v5, v6 ];
+
+export default deprecated;
